Advent of Code Day 18 solved par 1 fixed broken part and build process

+++ src/main/java/com/tmosest/competitiveprogramming/utils/string/StringUtil.java
@@ -110,7 +110,7 @@ public class StringUtil {
    */
   public static boolean isNumber(String str) {
     try {
-      Integer.valueOf(str);
+      Long.valueOf(str);
       return true;
     } catch (NumberFormatException e) {
       return false;
